Refactor Sidebar into three separate files
